Cover TypeScript's default exports with unit tests
